requirements.txt
pytest 
python-dotenv
selenium
webdriver-manager
requests
beautifulsoup4
allure-pytest
allure-python-commons
pyautogui
python-xlib
PyVirtualDisplay
python-10minutemail
minuteinbox-py